Unified generation outputs
