+++ v2/cmd/describeimage/main.go
@@ -102,11 +102,14 @@ func main() {
 	promptAndImages := append([]string{prompt}, images...)
 
 	logVerbose("[%s] Generating... ", oc.ModelName)
-	output, err := oc.GetOutput(promptAndImages...)
+	response, err := oc.GetOutput(promptAndImages...)
 	if err != nil {
 		fmt.Printf("error: %s\n", err)
 		os.Exit(1)
 	}
+
+	output := response.Response
+
 	logVerbose("OK\n")
 
 	if output == "" {
