+++ v2/stream.go
@@ -8,31 +8,6 @@ import (
 	"net/http"
 )
 
-// GenerateChatRequest represents the request payload for generating chat output
-type GenerateChatRequest struct {
-	Model    string         `json:"model"`
-	Messages []Message      `json:"messages,omitempty"`
-	Images   []string       `json:"images,omitempty"` // base64 encoded images
-	Stream   bool           `json:"stream"`
-	Tools    []Tool         `json:"tools,omitempty"`
-	Options  RequestOptions `json:"options,omitempty"`
-}
-
-// GenerateChatResponse represents the response data from the generate chat API call
-type GenerateChatResponse struct {
-	Model              string          `json:"model"`
-	CreatedAt          string          `json:"created_at"`
-	Message            MessageResponse `json:"message"`
-	DoneReason         string          `json:"done_reason"`
-	Done               bool            `json:"done"`
-	TotalDuration      int64           `json:"total_duration,omitempty"`
-	LoadDuration       int64           `json:"load_duration,omitempty"`
-	PromptEvalCount    int             `json:"prompt_eval_count,omitempty"`
-	PromptEvalDuration int64           `json:"prompt_eval_duration,omitempty"`
-	EvalCount          int             `json:"eval_count,omitempty"`
-	EvalDuration       int64           `json:"eval_duration,omitempty"`
-}
-
 // Message is a chat message
 type Message struct {
 	Role    string `json:"role"`
